Added fix for SOS: The Final Escape/Disaster Report/Zettai Zetsumei Toshi. This also includes new IOP patch functions.
The game suffered from a sound problem, which sometimes lead to instability.

There was a memo dated July 2000, which went along the lines of saying that the atick functions of modmidi and modhsyn should be called,
if related sequence data is stopepd and removed.

The explanation was that these modules only process sequence data when their atick functions are called.
If data is removed and their atick functions are not called before further callsto the library are made,
then the correct operation of the library cannot be guaranteed because the data would be dequeued.

So to correct this, the tick thread must be woken up after sequence data is dequeued.