Revert "Optional FMV fix (enabled by setting UNALIGNED_BUFFER_PATCH to 0)" Revert "I also like to commit in two-phases...:p"
Not only will there be no additional benefits, it will fail if the buffer is not aligned.
The original code already allows for performance when the buffer is aligned.

The original code was there to accommodate the restrictions imposed by the USBD module.
Due to hardware errata, buffers must be properly aligned. Or 63/64-byte frames be limited to 62, but not all devices are compatible with this other method.

The hardware errata seems to only concern the bulk-out pipe and not bulk-in.
However, the USBD module performs the alignment check on the pipe, as long as it was opened with sceUsbdOpenPipeAligned().

This reverts commit 6d5ae23990b1f911185f53c36d2d6f6dd0382332.
This reverts commit 27e4a47cd7e65d271dcfeb376e8a54f1c6c51a13.