Removed unofficial 480P and 1080P modes from GSM, added option for enabling kernel 576P fix (ROM v2.00 and earlier) and GCONT (RGB/YPbCbPrCr) setting will now follow the browser setting.
There exists a PSX with ROM v1.80, which seems to support 576P (0x53).
However, it has a lower version number than the late SCPH-50000 and SCPH-7000.
It shares the same GS as the SCPH-75000, but it is also shared with the SCPH-70000.
Hence I couldn't find a better way to identify whether the console supports 576P on its own.

There exists a PSX with v2.10, hence the minimum version can be set to v2.10 for now.