cdvdman: Fix for Samurai Warriors 2 and Warriors Orochi series. (#226)
Some games, such as Warriors Orochi (USA), calls sceCdReadIOPMem()
with a logical IOP RAM address (in this specific game, the passed
pointer is a KSEG0 - a.k.a. Kernel Mode, Segment 0 - logical IOP
RAM address) as its third parameter, which normally is not an issue,
but on slim consoles (SCPH-75000 series and newer), the game crashes
when being run from USB media. Fix this issue by ensuring that the
provided pointer to IOP RAM is a physical address.

Note that a similar measure is applied in cdvd_readee() (see cdvdfsv.c line 807):

http://github.com/ifcaro/Open-PS2-Loader/blob/master/modules/iopcore/cdvdfsv/cdvdfsv.c#L807

As a side note, this fix should apply for related koei games (Warriors
Orochi 2 and Samurai Warriors 2 series), as those games do the same
thing as Warriors Orochi regarding sceCdReadIOPMem().